Fix DB fetching bug
