+++ src/routes/api/events/+server.ts
@@ -3,10 +3,10 @@ import fs from 'fs/promises';
 import path from 'path';
 import toml from 'toml';
 import { Event, type Location } from '$lib/types/event';
-import sql from '$lib/db';
+import sql, { DB } from '$lib/db';
 
 export async function GET() {
-    Event.all().then((events) => {
-        return json(events);
-    })
+    const events = await DB.events();
+
+    return json(events);
 }
